Factor tests into api and response
