+++ tests/test_api.py
@@ -45,40 +45,7 @@ def test_metadata():
         'the Groningen Meaning Bank (GMB) dataset'
     assert metadata['license'] == 'Apache 2'
     assert metadata['type'] == 'Natural Language Processing'
-    assert 'max-named-entity-tagger' in metadata['source']
-
-
-def test_labels():
-    model_endpoint = 'http://localhost:5000/model/labels'
-
-    r = requests.get(url=model_endpoint)
-    assert r.status_code == 200
-
-    labels = r.json()
-    tags = labels['labels']
-    assert labels['count'] == 17
-    assert tags[0]['name'] == 'O'
-    assert tags[0]['id'] == '0'
-    assert tags[-1]['id'] == '16'
-    assert tags[-1]['name'] == 'I-ORG'
-
-
-def test_response():
-    model_endpoint = 'http://localhost:5000/model/predict'
-    text = 'John lives in Brussels and works for the EU'
-    test_json = {
-        "text": text
-    }
-    expected_tags = ["B-PER", "O", "O", "B-GEO", "O", "O", "O", "O", "B-ORG"]
-    expected_terms = ["John", "lives", "in", "Brussels", "and", "works", "for", "the", "EU"]
-
-    r = requests.post(url=model_endpoint, json=test_json)
-
-    assert r.status_code == 200
-    response = r.json()
-    assert response['status'] == 'ok'
-    assert response['prediction']['terms'] == expected_terms
-    assert response['prediction']['tags'] == expected_tags
+    assert metadata['source'] == 'https://developer.ibm.com/exchanges/models/all/max-named-entity-tagger/'
 
 
 if __name__ == '__main__':
